added some more comments and clean code
